# src/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "version": "1.0.0",
        "description": "API documentation for Account, Admin, Auth, Chat, and Usage endpoints."
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "Local development server"
        }
    ],
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "paths": {
        "/api/v1/account/change-name": {
            "patch": {
                "tags": [
                    "Account"
                ],
                "summary": "Change the logged-in user's name",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Name updated successfully."
                    }
                }
            }
        },
        "/api/v1/account/change-password": {
            "patch": {
                "tags": [
                    "Account"
                ],
                "summary": "Change the user's password",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "oldPassword": {
                                        "type": "string"
                                    },
                                    "newPassword": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password changed successfully."
                    }
                }
            }
        },
        "/api/v1/account/get-user": {
            "get": {
                "tags": [
                    "Account"
                ],
                "summary": "Get the currently authenticated user's details",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User info returned."
                    }
                }
            }
        },
        "/api/v1/admin/delete-user/{id}": {
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Permanently delete a user by ID",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User deleted successfully."
                    },
                    "404": {
                        "description": "User not found."
                    }
                }
            }
        },
        "/api/v1/admin/enpoint-stats": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Get endpoint usage/performance statistics",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Stats returned."
                    }
                }
            }
        },
        "/api/v1/admin/log-dashboard-open": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Log when an admin dashboard is opened",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Log entry stored."
                    }
                }
            }
        },
        "/api/v1/admin/me": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Get the current admin user's profile",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Admin info returned."
                    }
                }
            }
        },
        "/api/v1/admin/promote-user/{id}": {
            "put": {
                "tags": [
                    "Admin"
                ],
                "summary": "Promote a user by ID",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User promoted."
                    }
                }
            }
        },
        "/api/v1/admin/user-stats": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Get overall system user statistics",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Stats returned."
                    }
                }
            }
        },
        "/api/v1/admin/users": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "List all platform users",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User list returned."
                    }
                }
            }
        },
        "/api/v1/auth/{all}": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Authentication handler (catch-all route)",
                "parameters": [
                    {
                        "in": "path",
                        "name": "all",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "Sub-path for auth routing."
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Auth request processed."
                    }
                }
            }
        },
        "/api/v1/chat": {
            "post": {
                "tags": [
                    "Chat"
                ],
                "summary": "Send a chat request/message",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chat response generated."
                    }
                }
            }
        },
        "/api/v1/user-chat-usage": {
            "get": {
                "tags": [
                    "Usage"
                ],
                "summary": "Get the authenticated user's chat usage statistics",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usage stats returned."
                    }
                }
            }
        },
        "/api/v1/swagger": {
            "get": {
                "tags": [
                    "Internal"
                ],
                "summary": "Returns the OpenAPI specification (JSON)",
                "responses": {
                    "200": {
                        "description": "Swagger spec returned."
                    }
                }
            }
        },
        "/api/v1/swagger-ui": {
            "get": {
                "tags": [
                    "Internal"
                ],
                "summary": "Swagger UI documentation page",
                "responses": {
                    "200": {
                        "description": "HTML UI returned."
                    }
                }
            }
        }
    }
}